fix(fabric): Don't register cui packet type twice
Instead, the packet type and its codec is only registered when worldedit isn't installed. In the case that worldedit is installed, the encoder and decoder are added to the codec that is registered by worldedit for the clientside.

The custom payload system is not designed to have different clientside and serverside classes representing the same packet, so the whole thing isn't really type safe.